+++ confidence.go
@@ -45,7 +45,7 @@ type Confidence struct {
 	client       func(req *http.Request) (*http.Response, error)
 }
 
-func (confidence *Confidence) handleConfidence(quit chan struct{}, interval time.Duration, getDevice func() (interface{}, error)) {
+func (confidence *Confidence) handleConfidence(quit chan struct{}, interval time.Duration, getDevice func() interface{}) {
 	defer confidence.wg.Done()
 	t := time.NewTicker(interval)
 	defer t.Stop()
@@ -55,11 +55,7 @@ func (confidence *Confidence) handleConfidence(quit chan struct{}, interval time
 			return
 		case <-t.C:
 			go func() {
-				item, err := getDevice()
-				if err != nil {
-					return
-				}
-				device := item.(string)
+				device := getDevice().(string)
 				logging.Debug(confidence.logger).Log(logging.MessageKey(), "testing new device", "device", device)
 				confidence.measures.DeviceSize.Add(-1)
 				confidence.handleDevice(device)