+++ release_notes.md
@@ -1,2 +1,3 @@
 - bumped bascule version to v0.5.0
 - bumped webpa-common to v1.4.0
+- leverage capacityset in shuffle package